+++ apps/storybook/src/forms/RadioGroup.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from "@storybook/react";
 
-import { RadioGroup, RadioItem, Text } from "@optiaxiom/react";
+import { RadioGroup, RadioGroupItem, Text } from "@optiaxiom/react";
 
 const meta: Meta<typeof RadioGroup> = {
   component: RadioGroup,
@@ -13,9 +13,9 @@ type Story = StoryObj<typeof RadioGroup>;
 export const Basic: Story = {
   render: () => (
     <RadioGroup defaultValue="sample-1">
-      <RadioItem value="sample-1">Sample 1</RadioItem>
-      <RadioItem value="sample-2">Sample 2</RadioItem>
-      <RadioItem value="sample-3">Sample 3</RadioItem>
+      <RadioGroupItem value="sample-1">Sample 1</RadioGroupItem>
+      <RadioGroupItem value="sample-2">Sample 2</RadioGroupItem>
+      <RadioGroupItem value="sample-3">Sample 3</RadioGroupItem>
     </RadioGroup>
   ),
 };
@@ -23,9 +23,9 @@ export const Basic: Story = {
 export const Horizontal: Story = {
   render: () => (
     <RadioGroup defaultValue="sample-1" flexDirection="row">
-      <RadioItem value="sample-1">Sample 1</RadioItem>
-      <RadioItem value="sample-2">Sample 2</RadioItem>
-      <RadioItem value="sample-3">Sample 3</RadioItem>
+      <RadioGroupItem value="sample-1">Sample 1</RadioGroupItem>
+      <RadioGroupItem value="sample-2">Sample 2</RadioGroupItem>
+      <RadioGroupItem value="sample-3">Sample 3</RadioGroupItem>
     </RadioGroup>
   ),
 };
@@ -33,15 +33,15 @@ export const Horizontal: Story = {
 export const WithDescription: Story = {
   render: () => (
     <RadioGroup defaultValue="sample-1">
-      <RadioItem endDecorator={<Text>first</Text>} value="sample-1">
+      <RadioGroupItem endDecorator={<Text>first</Text>} value="sample-1">
         Sample 1
-      </RadioItem>
-      <RadioItem endDecorator={<Text>second</Text>} value="sample-2">
+      </RadioGroupItem>
+      <RadioGroupItem endDecorator={<Text>second</Text>} value="sample-2">
         Sample 2
-      </RadioItem>
-      <RadioItem endDecorator={<Text>third</Text>} value="sample-3">
+      </RadioGroupItem>
+      <RadioGroupItem endDecorator={<Text>third</Text>} value="sample-3">
         Sample 3
-      </RadioItem>
+      </RadioGroupItem>
     </RadioGroup>
   ),
 };
@@ -49,9 +49,9 @@ export const WithDescription: Story = {
 export const DisabledGroup: Story = {
   render: () => (
     <RadioGroup defaultValue="sample-1" disabled>
-      <RadioItem value="sample-1">Sample 1</RadioItem>
-      <RadioItem value="sample-2">Sample 2</RadioItem>
-      <RadioItem value="sample-3">Sample 3</RadioItem>
+      <RadioGroupItem value="sample-1">Sample 1</RadioGroupItem>
+      <RadioGroupItem value="sample-2">Sample 2</RadioGroupItem>
+      <RadioGroupItem value="sample-3">Sample 3</RadioGroupItem>
     </RadioGroup>
   ),
 };
@@ -59,14 +59,14 @@ export const DisabledGroup: Story = {
 export const MixedState: Story = {
   render: () => (
     <RadioGroup defaultValue="sample-1">
-      <RadioItem value="sample-1">Sample 1</RadioItem>
-      <RadioItem disabled value="sample-2">
+      <RadioGroupItem value="sample-1">Sample 1</RadioGroupItem>
+      <RadioGroupItem disabled value="sample-2">
         Sample 2
-      </RadioItem>
-      <RadioItem value="sample-3">Sample 3</RadioItem>
-      <RadioItem disabled value="sample-4">
+      </RadioGroupItem>
+      <RadioGroupItem value="sample-3">Sample 3</RadioGroupItem>
+      <RadioGroupItem disabled value="sample-4">
         Sample 4
-      </RadioItem>
+      </RadioGroupItem>
     </RadioGroup>
   ),
 };
@@ -74,9 +74,9 @@ export const MixedState: Story = {
 export const Readonly: Story = {
   render: () => (
     <RadioGroup defaultValue="sample-1" readonly>
-      <RadioItem value="sample-1">Sample 1</RadioItem>
-      <RadioItem value="sample-2">Sample 2</RadioItem>
-      <RadioItem value="sample-3">Sample 3</RadioItem>
+      <RadioGroupItem value="sample-1">Sample 1</RadioGroupItem>
+      <RadioGroupItem value="sample-2">Sample 2</RadioGroupItem>
+      <RadioGroupItem value="sample-3">Sample 3</RadioGroupItem>
     </RadioGroup>
   ),
 };
