+++ packages/react/src/radio-group-item/index.ts
@@ -1,0 +1,1 @@
+export * from "./RadioGroupItem";
