--- packages/react/src/radio-item/index.ts
@@ -1,1 +1,0 @@
-export * from "./RadioItem";
